Lots of basic functionality. Editable datamaps, syntax error detection, datamap inconsistency detection, autocomplete based on the datamap.
Still needs:
Actions to create datamap and rule files with prototypes / default contents.
Support for multiple datamaps / dealing with the datamap state hierarchy.
Support for datamap copy / pasting.
Actions to find rules that test a datamap attribute.
Error correction suggestions for errors / warnings.

+++ src/edu/umich/soar/editor/editors/datamap/actions/DeleteAttributeAction.java
@@ -1,11 +1,12 @@
 package edu.umich.soar.editor.editors.datamap.actions;
 
 import org.eclipse.jface.action.Action;
-import org.eclipse.jface.dialogs.InputDialog;
+import org.eclipse.jface.dialogs.MessageDialog;
+import org.eclipse.swt.graphics.Image;
+import org.eclipse.swt.widgets.Display;
 import org.eclipse.swt.widgets.Shell;
-import org.eclipse.ui.PlatformUI;
 
-import edu.umich.soar.editor.editors.datamap.Datamap.DatamapAttribute;
+import edu.umich.soar.editor.editors.datamap.DatamapAttribute;
 
 public class DeleteAttributeAction extends Action {
 
@@ -19,6 +20,19 @@ public class DeleteAttributeAction extends Action {
 	
 	@Override
 	public void run() {
-		attribute.delete();
+		String name = attribute.name;
+		Shell shell = Display.getDefault().getActiveShell();
+		String title = "Delete " + name + "?";
+		Image titleImage = null;
+		String message = "Delete " + name +"?";
+		int imageType = 0;
+		String[] buttons = { "OK", "Cancel" };
+		int defaultIndex = 0;
+		MessageDialog dialog = new MessageDialog(shell, title, titleImage, message, imageType, buttons, defaultIndex);
+		int result = dialog.open();
+		if (result == 0)
+		{
+			attribute.delete();
+		}
 	}
 }
